requirements.txt
# Server dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pyyaml>=6.0

# Test dependencies
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27.0   # also used by TestClient via starlette
